ml_model/requirements.txt
pyspark>=3.5.0
python-dotenv>=1.0.0 
pandas>=1.5.0